requirements.txt
confluent-kafka
environ-config
pydantic